Backend/proyectoflask/requirements.txt
Flask
firebase-admin
Pillow
numpy
werkzeug
flask_jwt_extended
pycuda
